requirements.txt
pyyaml~=5.4.1
google-cloud-translate==2.0.1